Split FileSystem.Windows into two additional files instead of one. Fix pinning.
