+++ src/System.IO.FileSystem/src/System.IO.FileSystem.csproj
@@ -19,6 +19,7 @@
     <Compile Include="System\IO\Enumeration\FileSystemEnumerableFactory.cs" />
     <Compile Include="System\IO\Enumeration\FileSystemEnumerator.cs" />
     <Compile Include="System\IO\Enumeration\FileSystemName.cs" />
+    <Compile Include="System\IO\FileSystem.DirectoryCreation.Windows.cs" />
     <Compile Include="System\IO\FileSystem.cs" />
     <Compile Include="System\IO\MatchCasing.cs" />
     <Compile Include="System\IO\MatchType.cs" />
@@ -55,7 +56,7 @@
     <Compile Include="System\IO\FileSystemInfo.Windows.cs" />
     <Compile Include="System\IO\Enumeration\FileSystemEntry.Windows.cs" />
     <Compile Include="System\IO\FileSystem.Windows.cs" />
-    <Compile Include="System\IO\FileSystem.AccessControl.Windows.cs" />
+    <Compile Include="System\IO\FileSystem.Attributes.Windows.cs" />
     <Compile Include="Microsoft\Win32\SafeHandles\SafeFindHandle.Windows.cs" />
     <Compile Include="$(CommonPath)\Interop\Windows\Interop.Libraries.cs">
       <Link>Common\Interop\Windows\Interop.Libraries.cs</Link>
